procore_seg/requirements.txt
biopython
fibos
h5py
minkowski-engine
numpy
pyyaml
requests
torch
tqdm
